Added Generic sensor type
Added a Generic sensor type to the air and sea guis. This sensor type uses the Hobo class to process data, so the input file must be in the style of Hobo data, or follow the Generic data file template that was created. One improvement was that a function has been added to csv_readers.py to detent the datetime string format used in the input data files. This function is currently only be used in the Hobo class.